neural_network/fred_handler.py
import numpy as np
from sklearn.manifold import MDS
from cryptography.hazmat.primitives.asymmetric import kyber, dilithium
from cryptography.hazmat.primitives import serialization
from cryptography.hazmat.backends import default_backend

class FredHandler:
    def __init__(self):
        self.data = None
        self.api = None
        self.security = None
        self.pqc_algorithms = {
            'kyber': kyber,
            'dilithium': dilithium
        }

    def load_data(self, data):
        self.data = data

    def set_api(self, api):
        self.api = api

    def set_security(self, security):
        self.security = security

    def integrate_pqc(self, algorithm_name):
        if algorithm_name in self.pqc_algorithms:
            algorithm = self.pqc_algorithms[algorithm_name]
            return algorithm
        else:
            raise ValueError("Unsupported PQC algorithm")

    def visualize_data(self):
        mds = MDS(n_components=2, dissimilarity='precomputed')
        transformed_data = mds.fit_transform(self.data)
        return transformed_data

    def perform_eigenvalue_analysis(self, matrix):
        eigenvalues, _ = np.linalg.eig(matrix)
        return eigenvalues

    def is_stable(self, matrix):
        eigenvalues = self.perform_eigenvalue_analysis(matrix)
        return np.all(np.real(eigenvalues) < 0)

    def optimize_parameters(self, params, matrix, learning_rate=0.01, max_iter=100):
        for _ in range(max_iter):
            eigenvalues = self.perform_eigenvalue_analysis(matrix)
            if self.is_stable(matrix):
                break
            params -= learning_rate * np.real(eigenvalues)
        return params

    def generate_random_seed(self):
        return np.random.randint(0, 2**32 - 1)

    def serialize_key(self, key):
        return key.public_bytes(
            encoding=serialization.Encoding.PEM,
            format=serialization.PublicFormat.SubjectPublicKeyInfo
        )

    def deserialize_key(self, key_bytes):
        return serialization.load_pem_public_key(key_bytes, backend=default_backend())

    def apply_neutrosophic_logic(self, data, truth, indeterminacy, falsity):
        return truth * data + indeterminacy * (1 - data) - falsity * data

    def filter_data(self, data, truth, indeterminacy, falsity):
        return self.apply_neutrosophic_logic(data, truth, indeterminacy, falsity)
