fred_handler/tests/test_fred_handler.py
import unittest
import numpy as np
from fred_handler.fred_handler import FredHandler
from fred_handler.visualization import QuantumDataVisualizer
from fred_handler.eigenvalue_analysis import EigenvalueAnalysis
from fred_handler.agent_based_modeling import AgentBasedModeling, QuantumAgent, QuantumEnvironment, NeutrosophicLogic, QuantumDataFiltration
from fred_handler.random_seed_manager import RandomSeedManager

class TestFredHandler(unittest.TestCase):

    def setUp(self):
        self.fred_handler = FredHandler()
        self.visualizer = QuantumDataVisualizer(data=np.random.rand(10, 10))
        self.eigenvalue_analysis = EigenvalueAnalysis(matrix=np.random.rand(5, 5))
        self.agents = [QuantumAgent(state=np.random.rand(4), strategy=lambda s, e: s)]
        self.environment = QuantumEnvironment(initial_conditions=np.random.rand(4))
        self.agent_based_modeling = AgentBasedModeling(agents=self.agents, environment=self.environment)
        self.neutrosophic_logic = NeutrosophicLogic(truth=0.7, indeterminacy=0.2, falsity=0.1)
        self.data_filtration = QuantumDataFiltration(neutrosophic_logic=self.neutrosophic_logic)
        self.random_seed_manager = RandomSeedManager(seed=42)

    def test_fred_handler_initialization(self):
        self.assertIsNotNone(self.fred_handler)

    def test_visualizer_initialization(self):
        self.assertIsNotNone(self.visualizer)

    def test_eigenvalue_analysis_initialization(self):
        self.assertIsNotNone(self.eigenvalue_analysis)

    def test_agent_based_modeling_initialization(self):
        self.assertIsNotNone(self.agent_based_modeling)

    def test_neutrosophic_logic_initialization(self):
        self.assertIsNotNone(self.neutrosophic_logic)

    def test_data_filtration_initialization(self):
        self.assertIsNotNone(self.data_filtration)

    def test_random_seed_manager_initialization(self):
        self.assertIsNotNone(self.random_seed_manager)

    def test_fred_handler_functionality(self):
        # Add tests for FredHandler functionality
        pass

    def test_visualizer_functionality(self):
        # Add tests for QuantumDataVisualizer functionality
        pass

    def test_eigenvalue_analysis_functionality(self):
        # Add tests for EigenvalueAnalysis functionality
        pass

    def test_agent_based_modeling_functionality(self):
        # Add tests for AgentBasedModeling functionality
        pass

    def test_neutrosophic_logic_functionality(self):
        # Add tests for NeutrosophicLogic functionality
        pass

    def test_data_filtration_functionality(self):
        # Add tests for QuantumDataFiltration functionality
        pass

    def test_random_seed_manager_functionality(self):
        # Add tests for RandomSeedManager functionality
        pass

if __name__ == '__main__':
    unittest.main()
